backend/requirements.txt
Flask>=2.0
Flask-CORS>=3.0
requests>=2.25
google-generativeai
python-dotenv
